Move postProcessByNameArgs to Nullables
